unit test refacto
