+++ phpunit.dist.legacy.xml
@@ -13,7 +13,7 @@
 
     <testsuites>
         <testsuite name="Alma PHP Client Unit Test Suite">
-            <directory>tests/Unit/{MYVERSION}</directory>
+            <directory>tests/Unit</directory>
         </testsuite>
         <testsuite name="Alma PHP Client Integration Test Suite">
             <directory>tests/Integration/{MYVERSION}</directory>
